Move CompositeUnit to separate file
